ya casi ya casi

+++ src/BibliotecaCalculadora.h
@@ -12,9 +12,9 @@
 
 int utn_getNumber(int* pResultado, char* mensaje, char* mensajeError, int minimo, int maximo, int reintentos);
 int multiplicar(int primerNumero, int segundoNumero, int* pResultado);
-float dividir(int primerNumero, int segundoNumero, float* pResultado);
+int dividir(int primerNumero, int segundoNumero, float* pResultado);
 int restar(int primerNumero, int segundoNumero, int* pResultado);
 int sumar(int primerNumero, int segundoNumero, int* pResultado);
-int factorial(int primerNumero, long int* pResultado);
+long int factorial(int primerNumero,long int* pResultado);
 
 #endif /* BIBLIOTECACALCULADORA_H_ */
